requirements/requirements.txt
opencv-python==3.4.18.65
Pillow==9.4.0
tensorflow==2.9.3
keras==2.9.3
numpy==1.24.1
Markdown==3.4.1
absl-py==1.4.0
Werkzeug==2.2.2
h5py==3.1.0
